pyproject.toml
[project]
name = "dtc-editor"
version = "0.3.0"
description = "DTC Editorial Engine vNext2: representation-first edit objects + deterministic apply + verification + bundle outputs + LLM rewrites + Google Docs export"
requires-python = ">=3.10"
dependencies = [
  "python-docx>=1.1.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.18.0"]
google = ["google-api-python-client>=2.100.0", "google-auth>=2.25.0"]
all = [
  "anthropic>=0.18.0",
  "google-api-python-client>=2.100.0",
  "google-auth>=2.25.0",
]

[project.scripts]
dtc-edit = "dtc_editor.cli:main"
dtc-edit-gui = "dtc_editor.gui:launch_gui"

[tool.pytest.ini_options]
testpaths = ["tests"]
